+++ CHANGES.md
@@ -5,6 +5,7 @@
 * Switched to semantic versioning.
 * Added offset and offsetSpeed options to allow for a scrolling/animated background effect.
 * Added color1, color2, color3, and color4 options to create gradients and psychedelic effects.
+* Added HTML5/GWT support.
 
 ### Version 3 ###